move component files to src/components
